# app/static/swagger.yaml
swagger: "2.0"
...
info:
  title: "Mechanic Database"
  description: "API for managing a mechanic shop system, including customers, service tickets, mechanics, and inventory."
  version: "1.0.0"

# API Host and Schemes
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"

# Security Definitions
securityDefinitions:
  bearerAuth:
    type: "apiKey"
    name: "Authorization"
    in: "header"

# API Endpoints
paths:
...
  /mechanics:
    get:
      tags:
        - "mechanics"
      summary: "List all mechanics"
      description: "Retrieve a list of all mechanics (requires mechanic authentication)."
      security:
        - bearerAuth: []
      responses:
        200:
          description: "Mechanics retrieved successfully"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/MechanicResponse"
    post:
      tags:
        - "mechanics"
      summary: "Create a new mechanic"
      description: "Add a new mechanic to the system."
      parameters:
        - in: "body"
          name: "body"
          description: "Mechanic information"
          required: true
          schema:
            $ref: "#/definitions/CreateMechanicPayload"
      responses:
        201:
          description: "Mechanic created successfully"
          schema:
            $ref: "#/definitions/MechanicResponse"
  /mechanics/login:
    post:
      tags:
        - "mechanics"
      summary: "Mechanic login"
      description: "Validates mechanic credentials and returns an authentication token."
      parameters:
        - in: "body"
          name: "body"
          description: "Login credentials"
          required: true
          schema:
            $ref: "#/definitions/LoginCredentials"
      responses:
        200:
          description: "Login successful"
          schema:
            $ref: "#/definitions/LoginResponse"
  /mechanics/{mechanic_id}:
    get:
      tags:
        - "mechanics"
      summary: "Get a mechanic"
      description: "Retrieve details of a specific mechanic by ID (requires mechanic authentication)."
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "mechanic_id"
          type: "integer"
          required: true
          description: "ID of the mechanic"
      responses:
        200:
          description: "Mechanic retrieved successfully"
          schema:
            $ref: "#/definitions/MechanicResponse"
    put:
      tags:
        - "mechanics"
      summary: "Update a mechanic"
      description: "Update an existing mechanic's information (requires mechanic authentication)."
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "mechanic_id"
          type: "integer"
          required: true
          description: "ID of the mechanic"
        - in: "body"
          name: "body"
          description: "Updated mechanic information"
          required: true
          schema:
            $ref: "#/definitions/CreateMechanicPayload"
      responses:
        200:
          description: "Mechanic updated successfully"
          schema:
            $ref: "#/definitions/MechanicResponse"
    delete:
      tags:
        - "mechanics"
      summary: "Delete a mechanic"
      description: "Remove a mechanic from the system (requires mechanic authentication)."
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "mechanic_id"
          type: "integer"
          required: true
          description: "ID of the mechanic"
      responses:
        200:
          description: "Mechanic deleted successfully"
          schema:
            $ref: "#/definitions/DeleteResponse"
  /mechanics/{mechanic_id}/service_tickets:
    get:
      tags:
        - "mechanics"
      summary: "List mechanic's service tickets"
      description: "Retrieve all service tickets assigned to a specific mechanic (requires mechanic authentication)."
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "mechanic_id"
          type: "integer"
          required: true
          description: "ID of the mechanic"
      responses:
        200:
          description: "Mechanic's service tickets retrieved successfully"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ServiceTicketResponse"
...
definitions:
...
  LoginCredentials:
    type: "object"
    properties:
      email:
        type: "string"
      password:
        type: "string"
    required:
      - email
      - password
  LoginResponse:
    type: "object"
    properties:
      message:
        type: "string"
      status:
        type: "string"
      token:
        type: "string"

  # Customer Schemas
...
  ServiceTicketResponse:
    type: "object"
    properties:
      customer_id:
        type: "integer"
      description:
        type: "string"
      labor_logs:
        type: "array"
        items:
          type: "object"
      mechanics:
        type: "array"
        items:
          type: "object"
          properties:
            email:
              type: "string"
            id:
              type: "integer"
            name:
              type: "string"
      service_date:
        type: "string"
      ticket_id:
        type: "integer"
      VIN:
        type: "string"
...
  CreateMechanicPayload:
    type: "object"
    properties:
      email:
        type: "string"
      name:
        type: "string"
      phone:
        type: "string"
      salary:
        type: "number"
      password:
        type: "string"
    required:
      - email
      - name
      - phone
      - salary
      - password
  MechanicResponse:
    type: "object"
    properties:
      email:
        type: "string"
      id:
        type: "integer"
      name:
        type: "string"
      phone:
        type: "string"
      salary:
        type: "number"
      password:
        type: "string"
...
  DeleteResponse:
    type: "object"
    properties:
      message:
        type: "string"
...